Completed the code and tests for the first approach
Made minor changes to edge_detector.py, notably the default values of a couple trackbars and the ability to set the kernal size to 0 (signifying no blurring).
Did some more experiments and added the images to 'edge test imgs', 'sketch-like test imgs', and 'blur & edge test imgs'.
Added a description text file to every folder of test images and added the original image as well as the original grayscale to three of those folders.

Future work: Now that the code is working, the next step is to speed up the run time of this first approach. Currently, the program takes a while to produce a result for large factors. It would also be interesting to see the outcome of changing how the image of edges is resized to the shape of the original image (by incrementally dividing the factor by 2 instead of decreasing it by 1, for example).